+++ packages/widgetbook/lib/src/next/addons/alignment_addon.dart
@@ -43,7 +43,7 @@ class AlignmentAddon extends ModeAddon<Alignment> {
   @override
   List<Field> get fields {
     return [
-      ListField<Alignment>(
+      ObjectDropdownField<Alignment>(
         name: 'alignment',
         initialValue: alignment,
         values: alignments.keys.toList(),