+++ packages/widgetbook/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - **BREAKING**: Set minimum SDK version to 3.7.0 & minimum Flutter version to 3.29.0. ([#1541](https://github.com/widgetbook/widgetbook/pull/1541))
 - **FEAT**: Add [`object.segmented`](https://docs.widgetbook.io/knobs/object/segmented) knob. ([#1478](https://github.com/widgetbook/widgetbook/pull/1478) - by [@Sourav-Sonkar](https://github.com/Sourav-Sonkar))
+- **REFACTOR**: Deprecate `ListField` in favor of `ObjectDropdownField`. ([#1541](https://github.com/widgetbook/widgetbook/pull/1541))
 
 ## 3.15.0